Added text chat function

+++ src/main/java/se/selimkose/labb3/Position.java
@@ -1,5 +1,5 @@
 package se.selimkose.labb3;
 
-public record Position(double x, double y) {
+public record Position(int x, int y) {
 
 }
